Refactored code, optimized imports and handled some checkstyle issues.
