+++ core/src/cg/group4/game_logic/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Package containing the game logic.
+ *
+ * @author Benjamin Los
+ */
+package cg.group4.game_logic;
