+++ core/src/cg/group4/game_logic/stroll/events/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Package containing all events regarding the strolls.
+ */
+package cg.group4.game_logic.stroll.events;
