+++ core/src/cg/group4/game_logic/stroll/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Package containing the mechanics of the stroll and events.
+ *
+ * @author Benjamin Los
+ */
+package cg.group4.game_logic.stroll;
