--- core/src/cg/group4/stroll/events/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Package containing all events regarding the strolls.
- */
-package cg.group4.stroll.events;
