--- core/src/cg/group4/stroll/package-info.java
@@ -1,8 +1,0 @@
-/**
- * @author TU Delf SID
- */
-/**
- * @author TU Delf SID
- *
- */
-package cg.group4.stroll;
