+++ core/src/cg/group4/util/sensors/package-info.java
@@ -3,8 +3,4 @@
  *
  * @author Jean de Leeuw
  */
-/**
- * @author Jean de Leeuw
- *
- */
 package cg.group4.util.sensors;
