+++ core/src/cg/group4/util/subscribe/Subject.java
@@ -18,6 +18,7 @@ public class Subject extends Observable {
 
     /**
      * Notifies the subscribers.
+     *
      * @param item Passes the object to its subscribers with the update.
      */
     public void update(Object item) {
