+++ core/src/cg/group4/util/subscribe/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Package containing the subscription mechanics for the observer pattern.
+ *
+ * @author Benjamin Los
+ */
+package cg.group4.util.subscribe;
