+++ core/src/cg/group4/util/timer/package-info.java
@@ -1,4 +1,7 @@
 /**
  * Package containing everything regarding timers.
+ *
+ * @author Benjamin Los
+ * @author Jurgen van Schagen
  */
 package cg.group4.util.timer;
