+++ core/src/cg/group4/view/screen/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Package containing the screens of the game.
+ *
+ * @author Benjamin Los
+ * @author Jurgen van Schagen
+ */
+package cg.group4.view.screen;
