+++ core/src/cg/group4/view/screen_mechanics/GameSkin.java
@@ -49,7 +49,7 @@ public class GameSkin extends Skin {
     /**
      * The height of the screen.
      */
-    float UI_HEIGHT;
+    protected float UI_HEIGHT;
 
     /**
      * Initializes the skin.
