+++ core/src/cg/group4/view/screen_mechanics/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Package containing the mechanics to display the screens.
+ *
+ * @author Benjamin Los
+ * @author Jurgen van Schagen
+ */
+package cg.group4.view.screen_mechanics;
